+++ libs/index.js
@@ -24,6 +24,10 @@ var _index11 = require("./v-dropdown/index");
 
 var _index12 = _interopRequireDefault(_index11);
 
+var _index13 = require("./v-draggable-table/index");
+
+var _index14 = _interopRequireDefault(_index13);
+
 function _interopRequireDefault(obj) { return obj && obj.__esModule ? obj : { default: obj }; }
 
 var install = function install(Vue) {
@@ -36,6 +40,7 @@ var install = function install(Vue) {
     Vue.component(_index8.default.name, _index8.default);
     Vue.component(_index10.default.name, _index10.default);
     Vue.component(_index12.default.name, _index12.default);
+    Vue.component(_index14.default.name, _index14.default);
 };
 
 if (typeof window !== 'undefined' && window.Vue) {
@@ -48,5 +53,6 @@ module.exports = {
     VCheckbox: _index6.default,
     VCheckboxGroup: _index8.default,
     VSelect: _index10.default,
-    VDropdown: _index12.default
+    VDropdown: _index12.default,
+    VDraggableTable: _index14.default
 };
