+++ packages/v-draggable-table/index.js
@@ -1,0 +1,7 @@
+import VDraggableTable from './src/draggable-table.vue';
+
+VDraggableTable.install = function(Vue) {
+    Vue.component(VDraggableTable.name, VDraggableTable);
+};
+
+export default VDraggableTable
